Delete empty translations on save.
This replaces the attempt in b5565e which attempted to remove blanks
before validations; something which ActiveRecord architecture does not
encourage.  This should be more robust.

The purpose of this is to keep the translations tables tidy.  If you
have a form that edits all translations via the locale-specific setters
you quickly end up with many unintentional blanks.  This also makes it
easy to query for objects missing certain translations.

For non-dependent translations an explicit destroy is required.